safari-600.7-branch/LayoutTests/http/tests/security/mixedContent/insecure-audio-video-in-main-frame-expected.txt

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-audio-video.html displayed insecure content from http://127.0.0.1:8080/resources/test.mp4.

CONSOLE MESSAGE: The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-audio-video.html displayed insecure content from http://127.0.0.1:8080/resources/test.mp4.

This test opens a window that loads insecure HTML5 audio and video. We should trigger a mixed content callback because the main frame in the window is HTTPS but is displaying insecure content.
